refactor: resolve circular dependency between SectionList components
