+++ src/app/newSettings/types/section.types.ts
@@ -1,0 +1,7 @@
+export interface NavSection {
+  section?: string;
+  subsection?: string;
+  text: string;
+  isSubsection?: boolean;
+  notificationsNumber: number;
+}
